Task 7: 完成角色权限模块实现
- 新增 sys_permission.go 权限服务（权限树、用户菜单、权限编码查询）
ollama run deepseek-r1:1.5b
- 新增 sys_permission.go 权限控制器（树列表、增删改查、用户权限）
- 更新 cmd.go 注册角色和权限路由（19个新接口）
- 修复 password.go rand.Int 调用错误
- 修复 Transaction 方法调用方式
- 修复 int/int64 类型转换问题
- 新增 response.go SuccessMsg/PageSuccess 方法
- 更新 tasks.md 标记 7.3/7.4/7.8 完成
- 更新 changelog.md 和未完成工作报告

+++ api-go/internal/service/sys_dict.go
@@ -82,7 +82,7 @@ func (s *sysDictServiceImpl) List(ctx context.Context, req *DictListReq) (*DictL
 
 	return &DictListRes{
 		Records:  dicts,
-		Total:    total,
+		Total:    int64(total),
 		Page:     req.Page,
 		PageSize: req.PageSize,
 	}, nil
